Inject `_TASKS` via constructor to `WorkflowsDeployment` instead of a global variable
Refactoring required for https://github.com/databrickslabs/ucx/pull/1209
